KINZ/pyproject.toml
[project]
name = "kinz"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "keras>=3.12.0",
    "matplotlib>=3.10.7",
    "numby>=0.2.0",
    "opencv-python>=4.12.0.88",
    "pillow>=12.0.0",
    "tensorflow>=2.20.0",
]
